Work on getting the race data into the API to complement the result data.
