+++ package.json
@@ -15,7 +15,8 @@
   "keywords": [
     "typescript",
     "koa",
-    "koa2"
+    "koa2",
+    "mongodb"
   ],
   "author": "Blair Garrett <blair.garrett1@gmail.com>",
   "license": "ISC",
@@ -32,7 +33,7 @@
     "@types/parse-ms": "^1.0.0",
     "@types/string-similarity": "^1.2.0",
     "axios": "^0.18.0",
-    "chalk": "^1.1.3",
+    "chalk": "1",
     "dotenv": "^6.0.0",
     "koa": "^2.0.0",
     "koa-bodyparser": "^3.1.0",
@@ -57,6 +58,7 @@
     "sinon-chai": "^2.8.0",
     "supertest": "^1.2.0",
     "supertest-as-promised": "^3.2.0",
+    "ts-jest": "^23.10.4",
     "ts-node": "^7.0.0",
     "tslint": "^3.15.1",
     "typescript": "^2.9.2"
